Fix IASA attacks not incrementing combo count
Fixed a "bug" where interrupting an aerial with the same aerial (most
commonly Marth fair->fair) wouldn't give it a unique move ID, causing it
to fail the combo count increment code.

This new injection adds a check to see if the current move is same as
the last one, and gives it a unique ID if so.