Corneria - move ship on stage initialization
I honestly cant believe this code is needed. Melee spawns the players at Y coordinate 300, but the stage starts at Y coordinate 0. Because of this, when i get the coordinate of the ground ID i want to spawn the players over, corneria is technically IN THE BLASTZONE. Therefore, i spawn the players in the blastzone and it loads state on death, so it infinitely kills the players and eventually crashes.

This code changes corneria's initialization code to move the ship, instead of waiting for the per frame think function to do it. This way the stage is in the right position when i
request the coordinates.