Fix event preloading
added preloading for custom events. this was completely borked before this, causing really long load times.

the reason for the load times was it was actually loading the original intended event stage first in full, then the one chosen for the event. the same went for character files.

now, all character files are loaded upon choosing them in the CSS and .ssm files are loaded upon leaving the CSS, just like in VS Mode. load times should now be on par with VS Mode.