stop all character SFX and crowd cheers when loading states + fix SDI event assertion
added code to stop all character SFX and crowd cheers when loading states. helps make it seem like its truly restoring a save state.

Fixed an assert when loading the SDI event as vanilla fox while in develop mode. Event mode code changes the CPU to orange fox but he isnt preloaded so it throws an assert. I should fix this by replacing the code that skips writing to preload table when the match starts with code that writes the selected characters and colors to preload table.